+++ History.md
@@ -3,6 +3,8 @@
 
 This incorporates all changes after 4.15.2 up to 4.15.4.
 
+  * remove:
+    - `res.send(status, body)` signature - use `res.status(status).send(body)`
   * deps: router@~1.3.1
     - deps: debug@2.6.8